% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcc_regions.R
\docType{data}
\name{bcc_regions}
\alias{bcc_regions}
\title{Regions of Birds of Conservation Concern}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4414 rows and 9 columns.
}
\usage{
bcc_regions
}
\description{
\code{bcc_regions} An \code{sf} object (polygon) of Birds of Conservation Concern
Regions.
}
\keyword{datasets}
