% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_basemap_data.R
\name{build_basemap_data}
\alias{build_basemap_data}
\title{Build base map data}
\usage{
build_basemap_data(admin_bndry, plan_area, target_crs = "EPSG:4326")
}
\arguments{
\item{admin_bndry}{An 'sf' object of the administrative Forest Service
boundary.}

\item{plan_area}{An 'sf' object of the plan area (Forest Service land).}

\item{target_crs}{The target coordinate reference system. The default is
EPSG:4326 (WGS 84).}
}
\value{
A list of \link[sf:sf]{sf} objects.
}
\description{
This is an alternative function to build base map data for the automated
reports. Spatial base map data for North and South America, the lower 48 US
states, and a user specified National Forest is returned in one concise
list.

\if{html}{\out{<div class="sourceCode">}}\preformatted{Continental and national scale data are acquired using the 
`rnaturalearth` package and Forest Service data provided from the user.
}\if{html}{\out{</div>}}
}
\examples{
message("Coming soon. To a theater near you...")
}
\seealso{
\code{\link[psoGIStools:read_edw_lyr]{read_edw_lyr()}}, \code{\link[rnaturalearth:ne_countries]{rnaturalearth::ne_countries()}},
\link[rnaturalearth:ne_states]{rnaturalearth::ne_states}, \code{\link[osmdata:osmdata_sf]{osmdata::osmdata_sf()}}
}
