% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_bbs_trend_narratives.R
\name{build_bbs_trend_narratives}
\alias{build_bbs_trend_narratives}
\title{Build Narratives for Breeding Bird Survey Trend Data}
\usage{
build_bbs_trend_narratives(region_codes, region_names)
}
\arguments{
\item{region_codes}{Vector of BBS region codes.}

\item{region_names}{Vector of BBS region names.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
This function builds the USGS Breeding Bird Survey (BBS) trend narratives for
bird species. The USGS breeding bird survey data can be found here:
https://www.sciencebase.gov/catalog/item/67aba702d34e329fb20457b1
}
\examples{
## Not run:

bbs_trends <-  build_bbs_trend_narratives(
  region_codes = c("IDA", "MON", "S10", "US1"),
  region_names = c("Idaho", "Montana",
                   "Northern Rockies Bird Conservation Region (S10)",
                   "the United States")
  )
 
## End(Not run)                     
}
