% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbcr_functions.R
\name{build_imbcr_spatial_data}
\alias{build_imbcr_spatial_data}
\title{Subset eligible species from IMBCR data.}
\usage{
build_imbcr_spatial_data(imbcr_data, spp_list)
}
\arguments{
\item{imbcr_data}{Spatial IMBCR data from \code{\link[=get_imbcr_data]{get_imbcr_data()}}.}

\item{spp_list}{Species list that includes taxon ID from \code{\link[=get_taxonomies]{get_taxonomies()}}.
This is the list that is used to subset the spatial data.}
}
\value{
An \link[sf:sf]{sf} object.
}
\description{
Subset eligible species from IMBCR data.
}
\details{
Additional details...
}
\examples{
\dontrun{
library("psoSppEvals")

# Read IMBCR data into R
mgmt_units <- c("Cimarron National Grassland", "Comanche National Grassland")
imbcr_dat <- get_imbcr_data(mgmt_units)

# Summarize data by species
spp_list <- build_imbcr_spp(imbcr_dat)
birds <- dplyr::filter(spp_list, family == "Corvidae")

# Subset spatial data
corvids <- build_imbcr_spatial_data(imbcr_dat, birds)
}
}
\seealso{
\code{\link[=get_imbcr_data]{get_imbcr_data()}}, \code{\link[=build_imbcr_spp]{build_imbcr_spp()}}, \code{\link[=get_taxonomies]{get_taxonomies()}}
}
