% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbcr_functions.R
\name{build_imbcr_spp}
\alias{build_imbcr_spp}
\title{Summarize IMBCR data by species}
\usage{
build_imbcr_spp(imbcr_data, dir_path = NULL)
}
\arguments{
\item{imbcr_data}{Spatial IMBCR data from \code{get_imbcr_data()}.}

\item{dir_path}{Path to the IMBCR data folder. Default is NULL. If NULL, data
will be read from the MPSG T-drive folder.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
This function summarizes the spatial IMBCR object from \code{get_imbcr_data()} by
species. The summary includes the number of records per species, minimum
and maximum year a species is observed, and taxonomy from the GBIF
backbone taxonomy. This function references restricted data on the
MPSG T-drive and will fail if the user is not on a Forest Service network
or the VDI and does not have access to the restricted data folder.
}
\examples{
## Not run:

library("mpsgSE")

# Read IMBCR data into R
mgmt_units <- c("Cimarron National Grassland", "Comanche National Grassland")
imbcr_dat <- get_imbcr_data(mgmt_units)

# Summarize data by species
build_imbcr_spp(imbcr_dat)

## End(Not run)
}
\seealso{
\code{\link[=get_imbcr_data]{get_imbcr_data()}}, \code{\link[=get_taxonomies]{get_taxonomies()}}
}
