% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucn_functions.R
\name{build_iucn_maps}
\alias{build_iucn_maps}
\title{Build IUCN Maps}
\usage{
build_iucn_maps(iucn_paths)
}
\arguments{
\item{iucn_paths}{}
}
\value{
An \link[sf:sf]{sf::sf} object.
}
\description{
This function reads spatial (polygon) range data from International Union for
Conservation of Nature (IUCN) into R. These shapefiles were downloaded
from IUCN (\url{https://www.iucnredlist.org/resources/spatial-data-download}),
and saved on the MPSG T-drive. This function will this function will fail
if you are not on a FS network or in the FS VDI. Use
\code{get_iucn_shp_paths()} to get the shapefile paths prior to running this
function.
}
\examples{
library(mpsgSE)
spp_list <- get_taxonomies(sp_list_ex)
bien_map_paths <- get_iucn_shp_paths(spp_list)
bien_maps <- build_iucn_maps(bien_map_paths)
}
\seealso{
\code{\link[=get_taxonomies]{get_taxonomies()}}, \code{\link[=get_iucn_shp_paths]{get_iucn_shp_paths()}}
}
