% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_map_source.R
\name{build_map_source}
\alias{build_map_source}
\title{Build Map Sources}
\usage{
build_map_source(spp_list, bien_maps, ebird_maps, iucn_maps)
}
\arguments{
\item{spp_list}{A species list with taxon ID's from \code{get_taxonomies()}.}

\item{bien_maps}{\code{sf} object from \code{download_bien_range_maps()}.}

\item{ebird_maps}{\code{sf} object from \code{download_ebird_range_maps()}.}

\item{iucn_maps}{\code{sf} object from \code{build_iucn_maps()}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
This function takes in the maps that were retrieved using
\code{download_bien_range_maps()},\code{download_ebird_range_maps()}, and
\code{build_iucn_maps()} and returns the data frame with taxonomic ID's,
taxonomies, and range map data sources.
}
\examples{
## Not run:

library("mpsgSE")
ebird_key <- "abcde12fghij34"
# Build speceis list
spp_list <- get_taxonomies(sp_list_ex)

# BIEN Maps
bien_range_maps <- download_bien_range_maps(spp_list, 
                                            file.path("data", "bien_maps"))

# eBird Maps
ebird_status <- download_ebird_status_maps(spp_list, 
                                           output_path = file.path("data/ebirdst"), 
                                           ebird_access_key = ebird_key)
ebird_maps <- download_ebird_range_maps(spp_list,
                                        output_path = file.path("data/ebirdst"),
                                        ebird_access_key = ebird_key)

# IUCN Maps
map_paths <- get_iucn_shp_paths(spp_list)
iucn_maps <- build_iucn_maps(map_paths)

# Build map source data frame
map_sources <- build_map_source(spp_list, bien_maps, ebird_maps, iucn_maps)

## End(Not run)                     
}
\seealso{
\code{\link[=get_taxonomies]{get_taxonomies()}},
}
