% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natureserve_functions.R
\name{build_ns_spp_list}
\alias{build_ns_spp_list}
\title{Build NatureServe Species List}
\usage{
build_ns_spp_list(ns_data)
}
\arguments{
\item{ns_data}{NatureServe data from \code{get_ns_state_list()}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
This function builds a species list with global and state NatureServe
rankings using data returned from \code{get_ns_state_list()}.
}
\examples{
## Not run:

library("mpsgSE")
ns_co <- get_ns_state_list("CO")
spp_ranks <- build_ns_spp_list(ns_co)

## End(Not run)                     
}
\seealso{
\code{\link[=get_ns_state_list]{get_ns_state_list()}}
}
