% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seinet_functions.R
\name{build_seinet_spp}
\alias{build_seinet_spp}
\title{Summarize SEINet data by species}
\usage{
build_seinet_spp(seinet_data, locale = TRUE)
}
\arguments{
\item{seinet_data}{Spatial SEINet data from \code{get_seinet_data()}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
This function summarizes the spatial SEINet object from \code{get_seinet_data()}
by species. The summary includes the number of records per species,
minimum and maximum year a species is observed, and the SEINet occurrence ID if
there are less than seven (7) observations.
}
\examples{
## Not run:

library("mpsgSE")

# Path to data folder
data_folder <- file.path("T:/path/to/project/directory", "data/SEINet")

# Pull data from existing SEINet query
sei_dat <- get_seinet_data(data_folder, crs = "NAD83")

# Summarize species
spp_list <- build_seinet_spp(sei_dat)

## End(Not run)
}
\seealso{
\code{\link[=get_seinet_data]{get_seinet_data()}}, \code{\link[=get_taxonomies]{get_taxonomies()}}
}
