% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{clip_fc}
\alias{clip_fc}
\title{Clip feature class to polygon}
\usage{
clip_fc(sf_lyr, sf_clip, locale)
}
\arguments{
\item{sf_lyr}{Spatial (\code{sf}) object to be clipped.}

\item{sf_clip}{Polygon (\code{sf}) object used to clip \code{sf_lyr} by.}

\item{locale}{Character. Short description of clipped layer, usually the
location (e.g., forest acronym or "Buffer").}
}
\value{
sf object
}
\description{
This function clips a \code{sf} object using \code{sf::st_intersection()}. First, this
function checks that the coordinate reference system (CRS) of the input
object is the same as the clipping object. If it is not, this function
transforms the CRS of the input object to the clipping object using
\code{sf::st_transform()}before clipping.
}
\examples{
## Not run:
devtools::install_github("fs-mschmidty/mpsgSE")
library("mpsgSE")

admin_bdy_sf <- read_lyr(lyr = "admin_bdy", 
                         dsn = file.path("T:/path/to/geodatabase"), 
                         crs = "NAD83")
                         
clip_fc(sf_lyr = gbif_sf, sf_clip = admin_bdy_sf, locale = "FS")

## End (Not run)
}
