% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natureserve_functions.R
\name{combine_natureserve_data}
\alias{combine_natureserve_data}
\title{Combine NatureServe Data}
\usage{
combine_natureserve_data(ns_list_1, ns_list_2)
}
\arguments{
\item{ns_list_1}{State list.}

\item{ns_list_2}{State list.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
This function combines the state NatureServe lists from the
\code{get_natureserve_state_data()} function in this pipeline.
}
