% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbcr_functions.R
\name{compile_imbcr_list}
\alias{compile_imbcr_list}
\title{Compile IMBCR list}
\usage{
compile_imbcr_list(imbcr_unit, imbcr_buff)
}
\arguments{
\item{imbcr_unit}{Spatial IMBCR data from \code{get_imbcr()} clipped to FS land.}

\item{imbcr_buff}{Spatial IMBCR data from \code{get_imbcr()} clipped to the 1-km
buffer of FS land.}
}
\value{
A tibble.
}
\description{
Compile a comprehensive species list from IMBCR occurrence records on Forest
Service (FS) land and a 1-km buffer of FS land. This function uses the
\code{imbcr_spp()} function on two clipped spatial objects from \code{get_imbcr()}.
}
\examples{
## Not run:

library("mpsgSE")

# Read spatial data into R
t_path <- file.path("T:/path/to/project/directory")
gdb_path <- file.path(t_path, "GIS_Data.gdb")
sf_fs <- read_fc(lyr = "PlanArea", dsn = gdb_path, crs = "NAD83")
sf_buff <- read_fc(lyr = "PlanArea_1kmBuffer", dsn = gdb_path, crs = "NAD83")

# Read IMBCR data into R
mgmt_units <- c("Cimarron National Grassland", "Comanche National Grassland")
imbcr_dat <- get_imbcr(mgmt_units)

# Clip to extents
unit_imbcr <- clip_fc(imbcr_dat, sf_fs)
buff_imbcr <- clip_fc(imbcr_dat, sf_buff)

compile_imbcr_list(unit_imbcr, buff_imbcr)

## End(Not run)                     
}
\seealso{
\code{\link[=get_imbcr]{get_imbcr()}}, \code{\link[=imbcr_spp]{imbcr_spp()}}, \code{\link[=clip_fc]{clip_fc()}}
}
