% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seinet.R
\name{compile_seinet_list}
\alias{compile_seinet_list}
\title{Compile SEINet list}
\usage{
compile_seinet_list(seinet_unit, seinet_buff)
}
\arguments{
\item{seinet_unit}{Spatial SEINet data from \code{get_seinet()} clipped to FS land.}

\item{seinet_buff}{Spatial SEINet data from \code{get_seinet()} clipped to the 1-km
buffer of FS land.}
}
\value{
A tibble.
}
\description{
Compile a comprehensive species list from SEINet occurrence records on Forest
Service (FS) land and a 1-km buffer of FS land. This function uses the
\code{seinet_spp()} function on two clipped spatial objects from
\code{get_seinet()}.
}
\examples{
## Not run:

library("mpsgSE")

# Read spatial data into R
t_path <- file.path("T:/path/to/project/directory")
gdb_path <- file.path(t_path, "GIS_Data.gdb")
sf_fs <- read_fc(lyr = "PlanArea", dsn = gdb_path, crs = "NAD83")
sf_buff <- read_fc(lyr = "PlanArea_1kmBuffer", dsn = gdb_path, crs = "NAD83")

# Pull data from existing SEINet query
sei_dat <- get_seinet(file.path(t_path, "data/SEINet"), crs = "NAD83")

# Clip to extents
unit_sei <- clip_fc(sei_dat, sf_fs)
buff_sei <- clip_fc(sei_dat, sf_buff)

# Summarize species
spp_list <- compile_seinet_list(unit_sei, buff_sei)

## End(Not run)                     
}
\seealso{
\code{\link[=get_seinet]{get_seinet()}}, \code{\link[=seinet_spp]{seinet_spp()}}, \code{\link[=clip_fc]{clip_fc()}}
}
