% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conservation_lists}
\alias{conservation_lists}
\title{Conservation lists in standardized format.  All lists have a taxon_id based
on the GBIF accepted taxonomy. Lists so far included are:
\itemize{
\item Colorado SWAP
\item Colorado T and E
\item USFS Region 2 Sensitive Species
}}
\format{
A data frame of 399 observations and 8 variables.
\describe{
\item{taxon_id}{MPSG taxon ID.}
\item{scientific_name}{Scientific name of species.}
\item{common_name}{Common name of species.}
\item{status_area}{Description.}
\item{status_authority}{Description.}
\item{status_all}{Description.}
\item{status_simple}{Description.}
\item{status_type}{Description.}
}
}
\usage{
conservation_lists
}
\description{
Conservation lists in standardized format.  All lists have a taxon_id based
on the GBIF accepted taxonomy. Lists so far included are:
\itemize{
\item Colorado SWAP
\item Colorado T and E
\item USFS Region 2 Sensitive Species
}
}
\keyword{datasets}
