% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{core_index}
\alias{core_index}
\title{Core Index data from the Breeding Bird Survey (USGS) 1966-2023}
\format{
A data frame of 988,146 observations and 6 variables.
\describe{
\item{aou}{BBS 5 digit codes that represent bird species.}
\item{region}{BBS region code.}
\item{year}{4 digit year.}
\item{index}{Annual indexes (Index) are relative abundance estimates for all years, estimated as yearly predicted abundances from the hierarchical model analysis.}
\item{ci_2.5_percenti}{2.5 percent confidence interval.}
\item{ci_97.5_percent}{97.5 percent confidence interval}
}
}
\source{
https://www.sciencebase.gov/catalog/item/67aba702d34e329fb20457b1
}
\usage{
core_index
}
\description{
Core Index data from the Breeding Bird Survey (USGS) 1966-2023
}
\keyword{datasets}
