% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{core_trend}
\alias{core_trend}
\title{Core Trend Data from the Breeding Bird Survey (USGS) 1966-2023}
\format{
A data frame of 17,037 observations and 16 variables.
\describe{
\item{credibility_code}{BBS model credibility code. R is a Red Indicator reflecting an impotrant deficiency in the model and Y is a Yellow Indicator reflecing a deficiency in the model.}
\item{sample_size_code}{Sample size component of the credibility code. R is based on less than 5 routes. Y is based on less than 14 routes. NA is based on at lease 14 routes.}
\item{precision_code}{Precision component to the credibility code. R is very imprecise. Y is quite imprecise. NA is precise.}
\item{abundance_code}{Abundance component to the credibility code. R is very low abundance. Y is low abundance. NA is abundant.}
\item{significance}{An indicator variable that takes on value "1" if zero falls outside the 95\% credible interval for the estimate and the trend estimate is > 0; "2" if zero falls outside the 95\% credible interval for the estimate and the trend estimate is < 0; and " " (no value) if zero falls within the 95\% credible interval. If the credible interval does not contain 0, the result could be judged significant.}
\item{aou}{BBS 5 digit codes that represent bird species.}
\item{region}{BBS region code.}
\item{species}{Common name of species.}
\item{region_name}{BBS region name.}
\item{model}{The statistical model used for the species.}
\item{n_routes}{The number of survey routes on which the species was encountered during the interval.}
\item{trend}{Trend estimates are calculated as the ratio of endpoints from the annual indices produced in the hierarchical model analysis and represent yearly percent change.}
\item{x2_5_percent_ci}{2.5 percent confidence interval.}
\item{x97_5_percent_ci}{9y.5 percent confidence interval.}
\item{relative_abundance}{The annual index for the region at the midpoint of the interval.}
\item{years}{Range of years that this trend covers.}
}
}
\source{
https://www.sciencebase.gov/catalog/item/67aba702d34e329fb20457b1
}
\usage{
core_trend
}
\description{
Core Trend Data from the Breeding Bird Survey (USGS) 1966-2023
}
\keyword{datasets}
