% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_taxon_ids.R
\name{correct_taxon_ids}
\alias{correct_taxon_ids}
\title{Correct known issues with taxon ID's}
\usage{
correct_taxon_ids(spp_list)
}
\arguments{
\item{spp_list}{A data frame with taxon ID's from \code{get_taxonomies()}.}
}
\value{
a data frame with corrected taxon ID's
}
\description{
Correct known issues with taxon ID's
}
\examples{
library(mpsgSE)

spp_list <- mpsgSE::sp_list_ex |>
  get_taxonomies() |> 
  correct_taxon_ids()
}
