% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natureserve_functions.R
\name{count_spp_by_hab}
\alias{count_spp_by_hab}
\title{Count Species by NatureServe Habitat Type}
\usage{
count_spp_by_hab(ns_habitats)
}
\arguments{
\item{ns_habitats}{Data frame from \code{get_ns_habitat()}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
This function creates a data frame that counts the number of species by
habitat in a data frame returned from \code{get_ns_habitat()}.
}
\examples{
## Not run:

library("mpsgSE")
ns_co <- get_ns_state_list("CO")
habitats <- get_ns_habitat(ns_co)
habitat_xwalk <- count_spp_by_hab(habitats)

## End(Not run)
}
\seealso{
\code{\link[=get_ns_habitat]{get_ns_habitat()}}
}
