% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_bien_range_maps.R
\name{download_bien_range_maps}
\alias{download_bien_range_maps}
\title{Download BIEN range maps}
\usage{
download_bien_range_maps(spp_list, output_path)
}
\arguments{
\item{spp_list}{List of species with taxonomies from \code{get_taxonomies()}.}

\item{output_path}{Directory path to save range map data to.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
This function downloads range maps from the Botanical Information and Ecology
Network (BIEN; \url{https://bien.nceas.ucsb.edu/bien/}) API for plants
identified as eligible for Species of Conservation Concern.
}
\examples{
library(mpsgSE)
spp_list <- get_taxonomies(sp_list_ex)
bien_range_maps <- download_bien_range_maps(spp_list,
                                            file.path("data", "bien_maps"))
}
\seealso{
\code{\link[=get_taxonomies]{get_taxonomies()}}
}
