% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_data.R
\docType{data}
\name{expanded_trend}
\alias{expanded_trend}
\title{Expanded Trend Data from the Breeding Bird Survey (USGS), 1993-2023}
\format{
\subsection{\code{expanded_trend}}{
}
}
\usage{
expanded_trend
}
\description{
Expanded Trend Data from the Breeding Bird Survey (USGS), 1993-2023
}
\keyword{datasets}
