% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fs_units}
\alias{fs_units}
\title{List of Forest Service Units}
\format{
A data frame
}
\source{
Forest Service EDW Rest Services, https://apps.fs.usda.gov/arcx/rest/services/EDW.
}
\usage{
fs_units
}
\description{
List of Forest Service Units
}
\keyword{datasets}
