% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_basemap_data.R
\name{get_basemap_data}
\alias{get_basemap_data}
\title{Get base map data}
\usage{
get_basemap_data(
  forest_name,
  admin_bndry = TRUE,
  plan_area = TRUE,
  target_crs = "EPSG:4326"
)
}
\arguments{
\item{forest_name}{Character. Name of national forest or grassland of
interest.}

\item{admin_bndry}{Optional. An \code{sf} object of the administrative boundary.
Default is TRUE. If TRUE, the \emph{EDW_ForestSystemBoundaries_01} layer will
be read in from the Forest Service ArcGIS REST service using
\code{\link[psoGIStools:read_edw_lyr]{psoGIStools::read_edw_lyr()}} and filtered using \emph{forest_name}.}

\item{plan_area}{Optional. An \code{sf} object of the administrative boundary.
Default is TRUE. If TRUE, the \emph{EDW_SurfaceOwnership_01} layer will
be read in from the Forest Service ArcGIS REST service using
\code{\link[psoGIStools:read_edw_lyr]{psoGIStools::read_edw_lyr()}} and clipped to the admin_bndry, then
filtered for Forest Service land.}

\item{target_crs}{Target coordinate reference system. Default is EPSG:4326
(WGS84).}
}
\value{
A list of \link[sf:sf]{sf} objects.
}
\description{
Pull spatial base map data for North and South America, the lower 48 US
states, and a user specified National Forest. Continental and national
scale data are acquired using the \code{rnaturalearth} package and Forest
Service data are acquired from Forest Service ArcGIS Rest Services
(https://apps.fs.usda.gov/arcx/rest/services/EDW) using \code{arcgislayers}
package.
}
\details{
\code{get_basemap_data} returns a list of spatial features used to produce
automated species evaluations.

Continental-scale data include 'americas', 'north_america', and 'l_48'. These
data are acquired using \code{\link[rnaturalearth:ne_countries]{rnaturalearth::ne_countries()}} and
\code{\link[rnaturalearth:ne_states]{rnaturalearth::ne_states()}} functions. These data are in the NAD83 CONUS
Albers (EPSG:5070).

National Forest-scale data include 'admin_bndry', 'plan_area',  'aoa',
and 'plan_area_doughnut'. 'admin_bndry' and 'plan_area' are acquired
using \code{\link[psoGIStools:read_edw_lyr]{read_edw_lyr()}}. 'aoa' and 'plan_area_doughnut' are derived using
the \link[sf:sf]{sf} package. 'aoa', or area of analysis, is a 4828 meter (3 mile)
buffer of 'admin_bndry' using \code{\link[sf:st_buffer]{sf::st_buffer()}}. 'plan_area_doughnut' is
a 1000 meter buffer of 'plan_area' with 'plan_area' erased using
\code{\link[sf:st_difference]{sf::st_difference()}}. These data are returned in the coordinate
reference system provided by the \code{crs} parameter.
}
\note{
Sometimes the connection to the Forest Service REST Service  fails. This will
throw the following error message: "\strong{Error:} tar_make() Status code:
500. Error: json". In most instances, the function can simply be executed
again to retrieve the data. Check the data servers if the error is
persistent.
}
\examples{
\dontrun{
library(psoSppEvals)
forest_name <- "Dixie National Forest"
basemap_data <- get_basemap_data(forest_name)
}
}
\seealso{
\code{\link[psoGIStools:read_edw_lyr]{read_edw_lyr()}}, \code{\link[rnaturalearth:ne_countries]{rnaturalearth::ne_countries()}},
\link[rnaturalearth:ne_states]{rnaturalearth::ne_states}, \code{\link[osmdata:osmdata_sf]{osmdata::osmdata_sf()}}
}
