% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcr_functions.R
\name{get_bc_regions}
\alias{get_bc_regions}
\title{Clip Birds of Conservation Regions to a polygon}
\usage{
get_bc_regions(sf_lyr)
}
\arguments{
\item{sf_lyr}{\link[sf:sf]{sf} polygon object.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
This function clips the Birds of Conservation Regions \link[sf:sf]{sf} object to a
polygon.
}
\examples{
\dontrun{
library("psoGIStools")
library("psoSppEvals")

# Pull the Administrative Boundary for the Dixie National Forest
adm_bdy <- read_edw_lyr("EDW_ForestSystemBoundaries_01") |> 
  dplyr::filter(region == "04" & forestnumber == "07")
get_bc_regions(adm_bdy)
}
}
