% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcc_functions.R
\name{get_bcc_regions}
\alias{get_bcc_regions}
\title{Clip Birds of Conservation Regions to a polygon}
\usage{
get_bcc_regions(sf_lyr)
}
\arguments{
\item{sf_lyr}{\link{sf} object.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
This function clips the Birds of Conservaiton Regions \link{sf} object to a
polygon
}
\examples{
adm_bdy <- mpsgSE::read_edw_lyr("EDW_ForestSystemBoundaries_01", layer = 1) |> 
  dplyr::filter(region == "04" & forestnumber == "07")
get_bcc_regions(adm_bdy)
}
