% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcr_functions.R
\name{get_bcr_list}
\alias{get_bcr_list}
\title{Get USGS Birds of Conservation Concern (BCC) breeding bird list for National
Forest or Grassland.}
\usage{
get_bcr_list(sf_lyr)
}
\arguments{
\item{sf_lyr}{An \code{sf} object of the Administrative Boundary of a National
Forest or Grassland.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
This function returns a data frame of bird species that occur in the Birds of
Conservation Regions a National Forest or Grassland overlaps.
}
\examples{
\dontrun{
library("psoGIStools")
library("psoSppEvals")

# Pull the Administrative Boundary for the Dixie National Forest
adm_bdy = read_edw_lyr("EDW_ForestSystemBoundaries_01") |>
  dplyr::filter(region == "04" & forestnumber == "07")

# Get list of birds of conservation concern
get_bcr_list(adm_bdy)
}
}
