% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_functions.R
\name{get_ebird_regional_stats}
\alias{get_ebird_regional_stats}
\title{Download regional trend statistics from eBird}
\usage{
get_ebird_regional_stats(spp_list, ebird_access_key)
}
\arguments{
\item{spp_list}{Species list with taxonomy from \code{\link[=get_taxonomies]{get_taxonomies()}}.}

\item{ebird_access_key}{eBird access key. Get a key by filling out the
request form at \url{https://ebird.org/st/request}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
This function downloads regional trend statistics from eBird using the
\code{ebirdst} package.
}
\details{
Additional details...
}
\examples{
## Not run:

library(mpsgSE)
spp_list <- get_taxonomies(sp_list_ex)
ebird_reg_status <- get_ebird_regional_stats(spp_list, ebird_access_key = ebird_key)

## End(Not run)                     
}
\seealso{
\code{\link[=get_taxonomies]{get_taxonomies()}}, \code{\link[=download_ebird_status_maps]{download_ebird_status_maps()}}
}
