% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_functions.R
\name{get_ebird_trends}
\alias{get_ebird_trends}
\title{Calculate abundance trends for bird species}
\usage{
get_ebird_trends(spp_list, output_path, ebird_access_key)
}
\arguments{
\item{spp_list}{Species list with taxonomy from \code{\link[=get_taxonomies]{get_taxonomies()}}.}

\item{output_path}{Directory path for trend data.}

\item{ebird_access_key}{eBird access key. Get a key by filling out the
request form at \url{https://ebird.org/st/request}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
This function calculates trends in bird species abundance from eBird data
using the \code{ebirdst} package.
}
\details{
Additional details...
}
\examples{
## Not run:

library("mpsgSE")

# Project directory path
t_path <- file.path("T:/path/to/project/directory/data")

# Pull data from existing GBIF query
gbif_dat <- get_gbif(gbif_key = '9999999-999999999999999', t_path)
# Summarize species
gbif_list <- gbif_spp(gbif_dat)
# Download status maps
gbif_birds <- get_ebird_trends(gbif_list, t_path, "a123b45c67d8")

## End(Not run)                     
}
\seealso{
\code{\link[=get_gbif]{get_gbif()}}, \code{\link[=gbif_spp]{gbif_spp()}}, \code{\link[=get_taxonomies]{get_taxonomies()}}
}
