% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imbcr.R
\name{get_imbcr}
\alias{get_imbcr}
\title{Read IMBCR Data into R.}
\usage{
get_imbcr(
  mgmt_unit,
  crs = NULL,
  dir_path = NULL,
  imbcr_file = NULL,
  ebird_file = NULL
)
}
\arguments{
\item{mgmt_unit}{The IMBCR management unit. Use \code{see_imbcr_mgmt_units()} to
see management unit names.}

\item{crs}{Target coordinate reference system (CRS). Either and
\code{\link[sf:st_crs]{sf::st_crs()}} object or accepted input string (e.g. "NAD83").
See \code{\link[sf:st_crs]{sf::st_crs()}} for more details. Default is NULL. If NULL,
resulting sf object CRS will be NAD83.}

\item{dir_path}{Path to the IMBCR data folder. Default is NULL. If NULL, data
will be read from the MPSG T-drive folder.}

\item{imbcr_file}{Name of IMBCR data file. Default is NULL. If NULL, the file
name in the MPSG T-drive will be used.}

\item{ebird_file}{Name of eBird taxa list file. Default is NULL. If NULL, the
file name in the MPSG T-drive will be used.}
}
\value{
An \link{sf} class object.
}
\description{
This function reads IMBCR data into R. The data are stored in an *.rda file
on the T-drive. The user can specify a file path if using a copy of the
*.rda file on your local machine.
}
\details{
You must be connected to the T-drive, e.g., through VPN, to use this function
unless you have the *.rda file on your local machine.
}
\examples{
## Not run:

devtools::install_github("fs-mschmidty/mpsgSE")
library("mpsgSE")

mgmt_units <- c("Cimarron National Grassland", "Comanche National Grassland")
get_imbcr(mgmt_units)

## End(Not run)                     
}
