% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbcr_functions.R
\name{get_imbcr_trends}
\alias{get_imbcr_trends}
\title{Read IMBCR trend data into R}
\usage{
get_imbcr_trends(bcc_region)
}
\arguments{
\item{bcc_region}{Character vector of BCC regions.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
This function read an *.RDS file into R and filters it by Birds of
Conservation Consern (BCC) region.
}
\examples{
## Not run:

library("mpsgSE")

regions <- c("UT-BCR9", "UT-BCR16", "UT")
imbcr_trends <- get_imbcr_trends(regions)

## End(Not run)
}
