% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucn_functions.R
\name{get_iucn_shp_paths}
\alias{get_iucn_shp_paths}
\title{Get paths to IUCN Shapefiles}
\usage{
get_iucn_shp_paths(eligible_list)
}
\arguments{
\item{eligible_list}{A species list with taxon ID's from \code{get_taxonomies()}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
This function reads in an *.RDS file of IUCN shapefile paths and filters it
to the eligible species list. The *.RDS file is on the MPSG T-drive, so
this function will fail if you are not on a FS network or in the FS VDI.
}
\examples{
## Not run:

library(mpsgSE)
spp_list <- get_taxonomies(sp_list_ex)
bien_map_paths <- get_iucn_shp_paths(spp_list)

## End(Not run)                     
}
\seealso{
\code{\link[=get_taxonomies]{get_taxonomies()}}
}
