% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natureserve_functions.R
\name{get_ns_habitat}
\alias{get_ns_habitat}
\title{Get NatureServe Habitats}
\usage{
get_ns_habitat(ns_state_list, spp_list)
}
\arguments{
\item{ns_state_list}{NatureServe State list from \code{\link[=get_ns_state_list]{get_ns_state_list()}}.}

\item{spp_list}{Species list with taxonomy from \code{\link[=get_taxonomies]{get_taxonomies()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
This function makes a call to the NatureServe API using the NatureServe State
List and returns all habitats associated with any given species.
}
\examples{
## Not run:

library("mpsgSE")
ns_co <- get_ns_state_list("CO")
habitats <- get_ns_habitat(ns_co)

## End(Not run)
}
