% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rfss_list.R
\name{get_rfss_list}
\alias{get_rfss_list}
\title{Get a Regional Forester's Sensitive Species List}
\usage{
get_rfss_list(fs_region)
}
\arguments{
\item{fs_region}{Character. Forest Service Region (e.g. "R1")}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
This funciton filter's the master Regional Forester's Sensitive Species List
saved as the \code{rfss} dataset in the \code{mpsgSE} package to a specified Forest
Service Region.
}
\examples{
install.packages(mpsgSE)
r1_ssl <- get_rfss_list("R1")
}
