% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seinet_functions.R
\name{get_seinet}
\alias{get_seinet}
\title{Read SEINet data into R}
\usage{
get_seinet(dir_path, crs = NULL)
}
\arguments{
\item{dir_path}{Path to SEINet data folder.}

\item{crs}{Target coordinate reference system (CRS). Either and
\code{\link[sf:st_crs]{sf::st_crs()}} object or accepted input string (e.g. "NAD83").
See \code{\link[sf:st_crs]{sf::st_crs()}} for more details. Default is NULL. If NULL,
resulting sf object CRS will be WGS84.}
}
\value{
An \code{sf} class object.
}
\description{
This function reads SEINet data in Darwin Core Archive format into R.
}
\details{
Additional details...
}
\examples{
## Not run:

library("mpsgSE")

# Path to data folder
data_folder <- file.path("T:/path/to/project/directory", "data/SEINet")

# Pull data from existing SEINet query
sei_dat <- get_seinet(data_folder, crs = "NAD83")

## End(Not run)                     
}
