% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seinet_functions.R
\name{get_seinet_data}
\alias{get_seinet_data}
\title{Read SEINet data into R}
\usage{
get_seinet_data(dir_path, crs = NULL, correct = TRUE)
}
\arguments{
\item{dir_path}{Path to SEINet data folder.}

\item{crs}{Target coordinate reference system (CRS). Either and
\code{\link[sf:st_crs]{sf::st_crs()}} object or accepted input string (e.g. "NAD83").
See \code{\link[sf:st_crs]{sf::st_crs()}} for more details. Default is NULL. If NULL,
resulting sf object CRS will be WGS84.}

\item{correct}{Logical. Run \code{correct_taxon_ids()} on data. Default is TRUE.}
}
\value{
An \code{sf} class object.
}
\description{
This function reads SEINet data in Darwin Core Archive format into R and
pulls taxon ID's and taxonomies from GBIF using \code{get_taxonomies()}
function.
}
\details{
SEINet data are available through an online data portal and need to be
downloaded manually. A polygon box, or well-known text (WKT) footprint,
is drawn around an area of interest to query species observations using
the Taxonomic Criteria search page
(\url{https://swbiodiversity.org/seinet/collections/harvestparams.php}).
The query results need to be manually downloaded in Darwin Core Archive
format and unzipped before using this function.
}
\examples{
## Not run:

library("mpsgSE")

# Path to data folder
data_folder <- file.path("T:/path/to/project/directory", "data/SEINet")

# Pull data from existing SEINet query
sei_dat <- get_seinet_data(data_folder, crs = "EPSG:26913")

## End(Not run)
}
\seealso{
\code{\link[=get_taxonomies]{get_taxonomies()}}
}
