% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_functions.R
\name{get_synonyms}
\alias{get_synonyms}
\title{Get Taxonomic Synonyms}
\usage{
get_synonyms(spp_list)
}
\arguments{
\item{spp_list}{Species list with taxon ID's from \code{get_taxonomies()}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
This function queries synonyms from the GBIF Backbone taxonomy. It will only
return synonyms for unique taxon ID's (i.e., duplicated taxon ID's will
not be queried).
}
\examples{
library(mpsgSE)
spp_data <- mpsgSE::sp_list_ex |>
  get_taxonomies('scientific_name')
get_synonyms(spp_data)
}
