% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imbcr.R
\name{imbcr_spp}
\alias{imbcr_spp}
\title{Summarize IMBCR data by species}
\usage{
imbcr_spp(imbcr_data)
}
\arguments{
\item{imbcr_data}{Spatial IMBCR data from \code{get_imbcr()}.}
}
\value{
A tibble.
}
\description{
This function summarizes the spatial IMBCR object from \code{get_imbcr()} by
species. The summary includes the number of records per species, minimum
and maximum year a species is observed, and the IMBCR record ID if there
are less than seven (7) observations. This function then verifies
taxonomy using the \code{get_taxonomies()} function.
}
\examples{
## Not run:

devtools::install_github("fs-mschmidty/mpsgSE")
library("mpsgSE")

# Read IMBCR data into R
mgmt_units <- c("Cimarron National Grassland", "Comanche National Grassland")
imbcr_dat <- get_imbcr(mgmt_units)

# Summarize data by species
imbcr_spp(imbcr_dat)

## End(Not run)                     
}
\seealso{
\code{\link[=get_imbcr]{get_imbcr()}}, \code{\link[=get_taxonomies]{get_taxonomies()}}
}
