% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rfss_list.R
\name{list_regions}
\alias{list_regions}
\title{List regions in rfss dataset}
\usage{
list_regions()
}
\value{
A \code{vector()}
}
\description{
This function lists the regions in the \code{rfss} dataset.
}
\examples{
install.packages(mpsgSE)
list_regions()
}
