% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_swap_list.R
\docType{data}
\name{mt_swap}
\alias{mt_swap}
\title{Montana SWAP List}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 154 rows and 12 columns.
}
\usage{
mt_swap
}
\description{
\code{mt_swap} Montana SWAP list from 2023.
}
\keyword{datasets}
