% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{name_corrections}
\alias{name_corrections}
\title{Taxon ID and Scientific Name corrections}
\format{
A data frame of 27 observations and 4 variables.
\describe{
\item{common_name}{Common species name.}
\item{errored_name}{Name that is not being queried in GBIF.}
\item{corrected_name}{A scientific name for the same speces in the GBIF backbome taxononmy.}
\item{taxon_id}{Taxon ID.}
}
}
\usage{
name_corrections
}
\description{
Manually generated through trial and error
}
\keyword{datasets}
