% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwi_functions.R
\name{pull_nwi_data}
\alias{pull_nwi_data}
\title{Read FWS National Wetlands Inventory}
\usage{
pull_nwi_data(plan_area_sf, gdb_path)
}
\arguments{
\item{plan_area_sf}{}
}
\value{
A list of \link[sf:sf]{sf::sf} objects: riparian and wetlands
}
\description{
Read FWS National Wetlands Inventory
}
\examples{
library(mpsgSE)
}
