% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{read_edw_lyr}
\alias{read_edw_lyr}
\title{Read spatial data from Forest Service ArcGIS REST Services}
\usage{
read_edw_lyr(map_name, layer = 0, crs = "EPSG:4326")
}
\arguments{
\item{map_name}{Character. Name of map layer.}

\item{layer}{Integer. Number of layer to read. Default is  zero (0).}

\item{crs}{Coordinate reference system (crs). Default is EPSG:4326 (WGS 84).}
}
\value{
An \link[sf:sf]{sf::sf} object or \link[terra:SpatRaster-class]{terra::SpatRaster}.
}
\description{
read_edw_lyr reads features from the Forest Service ArcGIS REST Services
Directory, https://apps.fs.usda.gov/arcx/rest/services/EDW, using the
\link{arcgislayers} package.
}
\examples{
library(mpsgSE)

# Administrative Boundary for the Dixie National Forest
admin_bndry <- read_edw_lyr("EDW_ForestSystemBoundaries_01", layer = 1) |> 
  dplyr::filter(forestname == "Dixie National Forest")

}
\seealso{
\code{\link[arcgislayers:arc_read]{arcgislayers::arc_read()}}, \code{\link[sf:st_transform]{sf::st_transform()}}
}
