% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{read_fc}
\alias{read_fc}
\title{Read feature class into R.}
\usage{
read_fc(lyr, dsn, crs = NULL)
}
\arguments{
\item{lyr}{Feature class name.}

\item{dsn}{Path to geodatabase that holds the feature class.}

\item{crs}{Target coordinate reference system (CRS). Either and
\code{sf::st_crs()} object or accepted input string for \code{sf::st_crs()} (e.g.
"WGS84" or "NAD83"). See \code{sf::st_crs()} for more details. Default is NULL.
If NULL, resulting sf object will not be transformed.}
}
\value{
sf object
}
\description{
This function uses the \code{sf} package to read a feature class into R from a
geodatabase using the \code{sf::read_sf()} function. It then checks that the
feature class is in the target coordinate reference system (CRS) and will
transform the feature to the target CRS if it is not.
}
\examples{
## Not run:

devtools::install_github("fs-mschmidty/mpsgSE")
library("mpsgSE")

read_fc(lyr = "feature_name", dsn = file.path("T:/path/to/geodatabase"), 
        crs = "NAD83")

## End (Not run)
}
\seealso{
\code{\link[sf:st_read]{sf::read_sf()}}, \code{\link[sf:st_crs]{sf::st_crs()}}
}
