% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rfss}
\alias{rfss}
\title{2024 Regional Forester's Sensitive Species List for Forest Service Regions 1-10}
\format{
A data frame of 4142 observations and 11 variables.
\describe{
\item{taxon_id}{MPSG taxon ID.}
\item{scientific_name}{Scientific name of species.}
\item{group_level2}{Taxonomic group}
\item{region}{Forest Service Region}
\item{rfssl}{Logical. Is the species on a Regional Forester's Sensitive Species List}
\item{region_num}{Forest Service Region number.}
\item{status_area}{Area label.}
\item{status_authority}{Authority label.}
\item{status_all}{Complete status label.}
\item{status_simple}{Simple status description.}
\item{status_type}{Status type}
}
}
\source{
Compiled from Regional Forester's Sensitive Species Lists as part of 2024 NOGA effort.
}
\usage{
rfss
}
\description{
2024 Regional Forester's Sensitive Species List for Forest Service Regions 1-10
}
\keyword{datasets}
