% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rfss_taxonomy}
\alias{rfss_taxonomy}
\title{Taxonomy for 2024 Regional Forester's Sensitive Species List for Forest Service Regions 1-10}
\format{
A data frame of 3746 observations and 12 variables.
\describe{
\item{taxon_id}{MPSG taxon ID.}
\item{scientific_name}{Scientific name of species.}
\item{kingdom}{Taxonomic Kingdom}
\item{phylum}{Taxonomic Phylum}
\item{class}{Taxonomic Class}
\item{order}{Taxonomic Order}
\item{family}{Taxonomic Family}
\item{genus}{Taxonomic Genus}
\item{species}{Taxonomic Species}
\item{subspecies}{Taxonomic Subspecies}
\item{form}{Taxonomic Form}
\item{variety}{Taxonomic Variety}
}
}
\source{
Compiled from Regional Forester's Sensitive Species Lists as part of 2024 NOGA effort. Taxonomy was generated using 'mpsgSE::get_taxonomies()'.
}
\usage{
rfss_taxonomy
}
\description{
Taxonomy for 2024 Regional Forester's Sensitive Species List for Forest Service Regions 1-10
}
\keyword{datasets}
