% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imbcr.R
\name{see_imbcr_mgmt_units}
\alias{see_imbcr_mgmt_units}
\title{View IMBCR Management Units}
\usage{
see_imbcr_mgmt_units(file_path = NULL)
}
\arguments{
\item{file_path}{Path to *.rda file.}
}
\value{
A vector.
}
\description{
This function reads IMBCR data into R. The data are stored in an *.rda file
on the T-drive. The user can specify a file path if using a copy of the
*.rda file on your local machine.
}
\details{
You must be connected to the T-drive, e.g., through VPN, to use this function
unless you have the *.rda file on your local machine.
}
\examples{
## Not run:

devtools::install_github("fs-mschmidty/mpsgSE")
library("mpsgSE")

see_imbcr_mgmt_units()

## End(Not run)                     
}
