% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seinet.R
\name{seinet_spp}
\alias{seinet_spp}
\title{Summarize SEINet data by species}
\usage{
seinet_spp(seinet_data, locale = TRUE)
}
\arguments{
\item{seinet_data}{Spatial SEINet data from \code{get_seinet()}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
This function summarizes the spatial SEINet object from \code{get_seinet()} by
species. The summary includes the number of records per species, minimum
and maximum year a species is observed, and the SEINet occurrence ID if
there are less than seven (7) observations. This function then verifies
taxonomy using the \code{get_taxonomies()} function.
}
\examples{
## Not run:

library("mpsgSE")

# Path to data folder
data_folder <- file.path("T:/path/to/project/directory", "data/SEINet")

# Pull data from existing SEINet query
sei_dat <- get_seinet(data_folder, crs = "NAD83")

# Summarize species
spp_list <- seinet_spp(sei_dat)

## End(Not run)                     
}
\seealso{
\code{\link[=get_seinet]{get_seinet()}}, \code{\link[=get_taxonomies]{get_taxonomies()}}
}
