% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_list_ex}
\alias{sp_list_ex}
\title{An example dataset of common names and scientific names of species.}
\format{
A data frame of 427 observations and 2 variables.
\describe{
\item{common_name}{Common name of species.}
\item{scientific_name}{Scientific name of species.}
}
}
\usage{
sp_list_ex
}
\description{
An example dataset of common names and scientific names of species.
}
\keyword{datasets}
