% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ut_blm_ss}
\alias{ut_blm_ss}
\title{Utah BLM Sensitive Species List}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 191 rows and 14 columns.
}
\usage{
ut_blm_ss
}
\description{
\code{ut_blm_ss} Utah BLM Sensitive Species list from 2025.
}
\keyword{datasets}
