% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut_swap_list.R
\docType{data}
\name{ut_swap}
\alias{ut_swap}
\title{Utah SWAP List}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 169 rows and 16 columns.
}
\usage{
ut_swap
}
\description{
\code{ut_swap} Utah SWAP list from 2025.
}
\keyword{datasets}
