% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcc_list.R
\docType{data}
\name{bcc_list}
\alias{bcc_list}
\title{US Fish & Wildlife Birds of Conservation Concern List}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 333 rows and 17 columns.
}
\usage{
bcc_list
}
\description{
\code{bcc_list} The US Fish & Wildlife Birds of Conservation Concern List from 2024.
}
\keyword{datasets}
