% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcc_regions}
\alias{bcc_regions}
\title{North American Bird Conservation Initiative Birds Conservation Regions}
\format{
An \code{sf} object (polygon) of 4414 observations and 9 variables.
\describe{
\item{bcr_label}{Birds Conservation Region codes.}
\item{bcr_label_name}{Birds Conservation Regions names.}
\item{name_en}{Birds Conservation Regions names in English.}
\item{name_fr}{Birds Conservation Regions names in French.}
\item{name_sp}{Birds Conservation Regions names in Spanish}
\item{globalid}{Polygon ID.}
\item{SHAPE_Length}{Polygon length.}
\item{SHAPE_Area}{Polygon area.}
\item{SHAPE}{Polygon geometry.}
}
}
\source{
https://www.birdscanada.org/bird-science/nabci-bird-conservation-regions
}
\usage{
bcc_regions
}
\description{
North American Bird Conservation Initiative Birds Conservation Regions
}
\keyword{datasets}
