% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_functions.R
\name{build_gbif_spatial_data}
\alias{build_gbif_spatial_data}
\title{Subset eligible species from GBIF data and reduce variables.}
\usage{
build_gbif_spatial_data(gbif_data, spp_list)
}
\arguments{
\item{gbif_data}{Spatial GBIF data from \code{\link[=get_gbif]{get_gbif()}}.}

\item{spp_list}{Species list that includes taxon ID from \code{\link[=get_taxonomies]{get_taxonomies()}}.
This is the list that is used to subset the spatial data.}
}
\value{
An \link[sf:sf]{sf::sf} object.
}
\description{
Subset eligible species from GBIF data and reduce variables.
}
\details{
Additional details...
}
\examples{
## Not run:

library("mpsgSE")

# Project directory path
t_path <- file.path("T:/path/to/project/directory/data")

# Pull data from existing GBIF query
gbif_dat <- get_gbif(gbif_key = '9999999-999999999999999', t_path = t_path)
# Summarize species
gbif_list <- gbif_spp(gbif_dat)
# Subset data
birds <- dplyr::filter(gbif_list, class == "Aves")
# Subset spatial data
gbif_birds <- build_gbif_spatial_data(gbif_dat, birds)

## End(Not run)
}
\seealso{
\code{\link[=get_gbif]{get_gbif()}}, \code{\link[=gbif_spp]{gbif_spp()}}, \code{\link[=get_taxonomies]{get_taxonomies()}}
}
