% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbcr_functions.R
\name{build_imbcr_spatial_data}
\alias{build_imbcr_spatial_data}
\title{Subset eligible species from IMBCR data.}
\usage{
build_imbcr_spatial_data(imbcr_data, imbcr_list, eligible_list)
}
\arguments{
\item{imbcr_data}{Spatial IMBCR data from \code{\link[=get_imbcr]{get_imbcr()}}.}

\item{imbcr_list}{List of IMBCR species from \code{\link[=imbcr_spp]{imbcr_spp()}}.}

\item{eligible_list}{Eligible species list that includes taxon ID from
\code{\link[=get_taxonomies]{get_taxonomies()}}. This is the list that is used to
subset the spatial data.}
}
\value{
An \link[sf:sf]{sf::sf} object.
}
\description{
Subset eligible species from IMBCR data.
}
\details{
Additional details...
}
\examples{
## Not run:

library("mpsgSE")

# Read IMBCR data into R
mgmt_units <- c("Cimarron National Grassland", "Comanche National Grassland")
imbcr_dat <- get_imbcr(mgmt_units)
# Summarize data by species
spp_list <- imbcr_spp(imbcr_dat)
birds <- dplyr::filter(spp_list, family == "Corvidae")
# Subset spatial data
corvids <- build_imbcr_spatial_data(imbcr_dat, spp_list, birds)

## End(Not run)
}
\seealso{
\code{\link[=get_imbcr]{get_imbcr()}}, \code{\link[=imbcr_spp]{imbcr_spp()}}, \code{\link[=get_taxonomies]{get_taxonomies()}}
}
