% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seinet_functions.R
\name{build_seinet_spatial_data}
\alias{build_seinet_spatial_data}
\title{Subset eligible species from SEINet data and reduce variables.}
\usage{
build_seinet_spatial_data(sei_data, spp_list)
}
\arguments{
\item{sei_data}{Spatial SEINet data from \code{\link[=get_seinet_data]{get_seinet_data()}}.}

\item{spp_list}{Species list that includes taxon ID from \code{\link[=get_taxonomies]{get_taxonomies()}}.
This is the list that is used to subset the spatial data.}
}
\value{
An \code{sf} class object.
}
\description{
Subset eligible species from SEINet data and reduce variables.
}
\details{
Additional details...
}
\examples{
## Not run:

library("mpsgSE")

# Path to data folder
data_folder <- file.path("T:/path/to/project/directory", "data/SEINet")

# Read SEINet data into R
sei_dat <- get_seinet_data(data_folder, crs = "NAD83")
# Create species list
spp_list <- sei_spp(sei_dat)
# Subset species list
birds <- dplyr::filter(class == "Aves")
# Subset spatial SEINet data
sei_birds <- build_seinet_spatial_data(sei_dat, birds)

## End(Not run)
}
\seealso{
\code{\link[=get_seinet_data]{get_seinet_data()}}, \code{\link[=build_seinet_spp]{build_seinet_spp()}}, \code{\link[=get_taxonomies]{get_taxonomies()}}
}
