% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbif.R
\name{compile_gbif_list}
\alias{compile_gbif_list}
\title{Compile GBIF list}
\usage{
compile_gbif_list(gbif_unit, gbif_buff)
}
\arguments{
\item{gbif_unit}{Spatial GBIF data from \code{get_gbif()} clipped to FS land.}

\item{gbif_buff}{Spatial GBIF data from \code{get_gbif()} clipped to the 1-km
buffer of FS land.}
}
\value{
A tibble.
}
\description{
Compile a comprehensive species list from GBIF occurrence records on Forest
Service (FS) land and a 1-km buffer of FS land. This function uses the
\code{gbif_spp()} function on two clipped spatial objects from \code{get_gbif()}.
}
\examples{
## Not run:

devtools::install_github("fs-mschmidty/mpsgSE")
library("mpsgSE")

# Read spatial data into R
t_path <- file.path("T:/path/to/project/directory")
gdb_path <- file.path(t_path, "GIS_Data.gdb")
sf_fs <- read_fc(lyr = "PlanArea", dsn = gdb_path, crs = "NAD83")
sf_buff <- read_fc(lyr = "PlanArea_1kmBuffer", dsn = gdb_path, crs = "NAD83")

# Pull data from existing GBIF query
gbif_dat <- get_gbif(gbif_key = '9999999-999999999999999', 
                     t_path = file.path(t_path, "data"))

# Clip to extents
unit_gbif <- clip_fc(gbif_sf, sf_fs)
buff_gbif <- clip_fc(gbif_sf, sf_buff)

# Summarize species
gbif_list <- compile_gbif_list(unit_gbif, buff_gbif)

## End(Not run)                     
}
\seealso{
\code{\link[=get_gbif]{get_gbif()}}, \code{\link[=gbif_spp]{gbif_spp()}}, \code{\link[=clip_fc]{clip_fc()}}
}
