% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservation_lists.R
\docType{data}
\name{conservation_lists}
\alias{conservation_lists}
\title{Conservation lists in standardised format.  All lists have a taxon_id based on the GBIF accepted taxonomy. Lists so far included are:
\itemize{
\item Colorado SWAP
\item Colorado T and E
\item USFS Region 2 Sensitive Species
}}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 399 rows and 8 columns.
}
\usage{
conservation_lists
}
\description{
Conservation lists in standardised format.  All lists have a taxon_id based on the GBIF accepted taxonomy. Lists so far included are:
\itemize{
\item Colorado SWAP
\item Colorado T and E
\item USFS Region 2 Sensitive Species
}
}
\keyword{datasets}
