% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_data.R
\docType{data}
\name{core_index}
\alias{core_index}
\title{Core Index data from the Breeding Bird Survey (USGS) 1966-2023}
\format{
\subsection{\code{core_index}}{
}
}
\usage{
core_index
}
\description{
Core Index data from the Breeding Bird Survey (USGS) 1966-2023
}
\keyword{datasets}
