% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_data.R
\docType{data}
\name{core_trend}
\alias{core_trend}
\title{Core Trend Data from the Breeding Bird Survey (USGS) 1966-2023}
\format{
\subsection{\code{core_trend}}{
}
}
\usage{
core_trend
}
\description{
Core Trend Data from the Breeding Bird Survey (USGS) 1966-2023
}
\keyword{datasets}
