% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_functions.R
\name{download_ebird_range_maps}
\alias{download_ebird_range_maps}
\title{Download eBird range maps}
\usage{
download_ebird_range_maps(spp_list, output_path, ebird_access_key)
}
\arguments{
\item{spp_list}{Species list with taxonomy from \code{\link[=get_taxonomies]{get_taxonomies()}}.}

\item{output_path}{Directory path for range maps.}

\item{ebird_access_key}{eBird access key. Get a key by filling out the
request form at \url{https://ebird.org/st/request}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
This function downloads range maps from eBird using the \code{ebirdst} package.
}
\details{
Additional details...
}
\examples{
## Not run:

library(mpsgSE)
spp_list <- get_taxonomies(sp_list_ex)
ebird_status <- download_ebird_status_maps(spp_list, 
                                           output_path = file.path("data/ebirdst"), 
                                           ebird_access_key = ebird_key)
ebird_maps <- download_ebird_range_maps(spp_list,
                                        output_path = file.path("data/ebirdst"),
                                        ebird_access_key = ebird_key)

## End(Not run)                     
}
\seealso{
\code{\link[=get_taxonomies]{get_taxonomies()}}, \code{\link[=download_ebird_status_maps]{download_ebird_status_maps()}}
}
