% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_functions.R
\name{download_ebird_status_maps}
\alias{download_ebird_status_maps}
\title{Download eBird status maps}
\usage{
download_ebird_status_maps(spp_list, output_path, ebird_access_key)
}
\arguments{
\item{spp_list}{Species list with taxonomy from \code{\link[=get_taxonomies]{get_taxonomies()}}.}

\item{output_path}{Directory path for status maps.}

\item{ebird_access_key}{eBird access key. Get a key by filling out the
request form at \url{https://ebird.org/st/request}.}
}
\value{
\code{\link[tibble:lst]{tibble::lst()}}
}
\description{
This function downloads status maps from eBird using the \code{ebirdst} package.
}
\details{
Additional details...
}
\examples{
## Not run:

library(mpsgSE)
spp_list <- get_taxonomies(sp_list_ex)
ebird_status <- download_ebird_status_maps(spp_list, 
                                           output_path = file.path("data/ebirdst"), 
                                           ebird_access_key = ebird_key)

## End(Not run)                     
}
\seealso{
\code{\link[=get_taxonomies]{get_taxonomies()}}
}
