% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_data.R
\docType{data}
\name{expanded_index}
\alias{expanded_index}
\title{Expanded Index data from the Breeding Bird Survey (USGS) 1993-2022}
\format{
\subsection{\code{expanded_index}}{
}
}
\usage{
expanded_index
}
\description{
Expanded Index data from the Breeding Bird Survey (USGS) 1993-2022
}
\keyword{datasets}
