% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_units.R
\docType{data}
\name{fs_units}
\alias{fs_units}
\title{List of Forest Service Units}
\format{
An object of class \code{data.frame} with 112 rows and 6 columns.
}
\usage{
fs_units
}
\description{
\code{fs_units} List of USDA Forest Service units from the Forest Service Rest
Services, https://apps.fs.usda.gov/arcx/rest/services/EDW.
}
\keyword{datasets}
