% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbif.R
\name{gbif_spatial}
\alias{gbif_spatial}
\title{Convert GBIF data frame to an sf object}
\usage{
gbif_spatial(gbif_dat, crs = NULL)
}
\arguments{
\item{gbif_dat}{GBIF data frame from \code{get_gbif()}.}

\item{crs}{Target coordinate reference system (CRS). Either and
\code{sf::st_crs()} object or accepted input string for
\code{sf::st_crs()} (e.g. "WGS84" or "NAD83"). See \code{sf::st_crs()}
for more details. Default is NULL. If NULL, resulting sf
object CRS will be WGS84.}
}
\value{
An sf object.
}
\description{
This function converts the output from \code{get_gbif()} to a spatial (\code{sf})
object using \code{sf::st_as_sf()}. This function will also transform the data
to a target coordinate reference system.
}
\examples{
## Not run:

devtools::install_github("fs-mschmidty/mpsgSE")
library("mpsgSE")

# Project directory path
t_path <- file.path("T:/path/to/project/directory")

# Pull data from existing GBIF query
gbif_dat <- get_gbif(gbif_key = '9999999-999999999999999', 
                     t_path = file.path(t_path, "data"))

# Convert to spatial object
gbif_sf <- gbif_spatial(gbif_dat, "NAD83")

## End(Not run)                     
}
\seealso{
\code{\link[=get_gbif]{get_gbif()}}, \code{\link[sf:st_as_sf]{sf::st_as_sf()}}, \code{\link[sf:st_crs]{sf::st_crs()}}
}
