% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_functions.R
\name{gbif_spp}
\alias{gbif_spp}
\title{Summarize GBIF data by species}
\usage{
gbif_spp(gbif_data, locale = TRUE, correct = TRUE)
}
\arguments{
\item{gbif_data}{Spatial GBIF data from \code{get_gbif()}.}

\item{locale}{Logical. Location description of data. E.g., unit acronym or "Buffer"}

\item{correct}{Logical. Run \code{correct_taxon_ids()} on data. Default is TRUE.}
}
\value{
A tibble.
}
\description{
This function summarizes the spatial GBIF object from \code{get_gbif()} by
species. Currently this function only works when
\code{get_gbif(..., process_data = TRUE)}. The summary includes the number of
records per species, minimum and maximum year a species is observed, and
the GBIF occurrence ID if there are less than seven (7) observations.
This function then verifies taxonomy using the \code{get_taxonomies()}
function.
}
\examples{
## Not run:

library("mpsgSE")

# Project directory path
t_path <- file.path("T:/path/to/project/directory")

# Pull data from existing GBIF query
gbif_dat <- get_gbif(gbif_key = '9999999-999999999999999',
                     t_path = file.path(t_path, "data"))

# Summarize species
gbif_list <- gbif_spp(gbif_dat)

## End(Not run)
}
\seealso{
\code{\link[=get_gbif]{get_gbif()}}, \code{\link[=get_taxonomies]{get_taxonomies()}}, \code{\link[=correct_taxon_ids]{correct_taxon_ids()}}
}
