% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{get_basemap_data}
\alias{get_basemap_data}
\title{Get base map data}
\usage{
get_basemap_data(
  states,
  region_number,
  forest_number,
  forest_name,
  crs = "EPSG:26912"
)
}
\arguments{
\item{states}{A list of state names or abbreviations.}

\item{region_number}{The Forest Service Region number}

\item{forest_number}{The Forest Service Forest number.}

\item{forest_name}{The Name of the National Forest.}

\item{crs}{The target coordinate reference system. The default is EPSG:26912
(NAD83 UTM Zone 12).}
}
\value{
A list.
}
\description{
This function pull spatial base map data for North and South America, the
lower 48 US states, and a user specified National Forest. Continental and
national scale data are acquired uisng the \link{rnaturalearth} package and
Forest Serivce data are acquired from Forest Service EDW Rest Services
(https://apps.fs.usda.gov/arcx/rest/services/EDW) using \link{arcgislayers}
package. Roads data are acquired using the \link{osmdata} package.
}
\examples{
library(mpsgSE)
states <- c("Utah", "Nevada", "New Mexico")
region_number <- "04"
forest_number <- "07"
forest_name <- "Dixie National Forest"
basemap_data <- get_basemap_data(states, region_number, forest_number, 
                                 forest_name)
}
