% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{get_basemap_data}
\alias{get_basemap_data}
\title{Get base map data}
\usage{
get_basemap_data(
  states,
  region_number,
  forest_number,
  forest_name,
  crs = "EPSG:26912"
)
}
\arguments{
\item{states}{A list of state names or abbreviations.}

\item{region_number}{The Forest Service Region number}

\item{forest_number}{The Forest Service Forest number.}

\item{forest_name}{The Name of the National Forest.}

\item{crs}{The target coordinate reference system. The default is EPSG:26912
(NAD83 UTM Zone 12).}
}
\value{
A list of \link{sf} objects.
}
\description{
Pull spatial base map data for North and South America, the lower 48 US
states, and a user specified National Forest. Continental and national
scale data are acquired using the \link{rnaturalearth} package and Forest
Service data are acquired from Forest Service EDW Rest Services
(https://apps.fs.usda.gov/arcx/rest/services/EDW) using \link{arcgislayers}
package. Roads data are acquired using the \link{osmdata} package.
}
\details{
\code{get_basemap_data} returns a list of spatial features used to produce
automated species evaluations.

Continental-scale data include 'americas', 'north_america', and 'l_48'. These
data are acquired using \code{\link[rnaturalearth:ne_countries]{rnaturalearth::ne_countries()}} and
\code{\link[rnaturalearth:ne_states]{rnaturalearth::ne_states()}} functions. These data are in the NAD83 CONUS
Albers (EPSG:5070).

National Forest-scale data include 'admin_bndry', 'plan_area', 'districts',
'buffer', and 'aoa_bbox.' 'admin_bndry', 'plan_area', and 'districts' are
acquired using \code{\link[=read_edw_lyr]{read_edw_lyr()}}. 'buffer' and 'aoa_bbox' are derived
using the \link{sf} package. These data are returned in the coordinate
reference system provided by the \code{crs} parameter.

Roads data include 'roads'. These data are acquired using the \link{osmdata}
package, primarily \code{\link[osmdata:getbb]{osmdata::getbb()}}, \code{\link[osmdata:opq]{osmdata::opq()}}, and
\code{\link[osmdata:osmdata_sf]{osmdata::osmdata_sf()}}. These data are returned in the coordinate
reference system provided by the \code{crs} parameter.
}
\note{
Sometimes the connection to the Forest Service REST Service or \link{osmdata}
fails. This will throw the following error message: "\strong{Error:}
tar_make() Status code: 500. Error: json". In most instances, the
function can simply be executed again to retrieve the data. Check the
data servers if the error is persistent.
}
\examples{
library(mpsgSE)
states <- c("Utah", "Nevada", "New Mexico")
region_number <- "04"
forest_number <- "07"
forest_name <- "Dixie National Forest"
basemap_data <- get_basemap_data(states, region_number, forest_number, 
                                 forest_name)
}
\seealso{
\code{\link[=read_edw_lyr]{read_edw_lyr()}}, \code{\link[rnaturalearth:ne_countries]{rnaturalearth::ne_countries()}},
\link[rnaturalearth:ne_states]{rnaturalearth::ne_states}, \code{\link[osmdata:osmdata_sf]{osmdata::osmdata_sf()}}
}
