% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcc_functions.R
\name{get_bcc_list}
\alias{get_bcc_list}
\title{Get USGS Birds of Conservation Concern (BCC) breeding bird list for National
Forest or Grassland.}
\usage{
get_bcc_list(sf_lyr)
}
\arguments{
\item{sf_lyr}{An \code{sf} object of the Administrative Boundary of a Natioal
Forest or Grassland.}
}
\value{
A data frame.
}
\description{
This function returns a data frame of bird species that occur in the Birds of
Conservation Regions a National Forest or Grassland overlaps.
}
\examples{
# Pull the Administrative Boundary for the Dixie National Forest
adm_bdy = mpsgSE::read_edw_lyr("EDW_ForestSystemBoundaries_01", layer = 1) |>
  dplyr::filter(region == "04" & forestnumber == "07")

# Get list of birds of conservation concern
mpsgSE::get_bcc_list(adm_bdy)
}
