% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbif.R
\name{get_gbif}
\alias{get_gbif}
\title{Query GBIF species occurrence records}
\usage{
get_gbif(
  gbif_key,
  t_path,
  aoa_wkt = NULL,
  gbif_user = NULL,
  gbif_pwd = NULL,
  gbif_email = NULL,
  gbif_format = "DWCA",
  process_data = TRUE
)
}
\arguments{
\item{gbif_key}{Character. The 22-digit GFIB key including hyphen for the
data set. Use "new" for new GBIF queries.}

\item{t_path}{Character. The directory path where the GBIF data package is or
will be stored.}

\item{aoa_wkt}{Vector. Well-known text (wtk) string for the area of analysis.
Use \code{wkt_string} to generate the wtk string. This is
required when \code{gbif_key} is set to \strong{new}. Default is
NULL.}

\item{gbif_user}{Character. Your GBIF user name. This is required when
\code{gbif_key} is set to \strong{new}. Default is NULL.}

\item{gbif_pwd}{Character. Your GBIF password. This is required when
\code{gbif_key} is set to \strong{new}. Default is NULL.}

\item{gbif_email}{Character. Your GBIF email address. This is required when
\code{gbif_key} is set to \strong{new}. Default is NULL.}

\item{gbif_format}{Character. The format of the data returned from GBIF.
Default is Darwin-Core Achrive (DWAC). See
\code{rgbif::occ_download} for more details.}

\item{process_data}{Logical (TRUE/FALSE). Are data to be processed after
reading them into R. Default is TRUE. The processing
step 1) filters the data for species, subspecies, and
varieties, 2) filters the data for present records,
3) filters against fossil records, 4) assembles clean
scientific names, without authority, from the genus,
specific epithet, and infraspecific epithet
variables, and 5) attempts to parse dates, day of
year, and year values.}
}
\value{
A tibble
}
\description{
Query GBIF species occurrence records
}
