% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natureserve_functions.R
\name{get_ns_state_list}
\alias{get_ns_state_list}
\title{Import NatureServe Species Data}
\usage{
get_ns_state_list(state, taxonomy = TRUE)
}
\arguments{
\item{state}{A character with state code, example: "CO".}

\item{taxonomy}{TRUE of FALSE.  TRUE if you want to return the data with
taxonomic information included.  This can take a long
time for states with large lists.}
}
\value{
A \code{\link[=list]{list()}}
}
\description{
This function imports NatureServe species data for a given state into R.
}
\examples{
## Not run:

library("mpsgSE")
co_ns_data <- get_ns_state_list("CO")

## End(Not run)
}
