% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natureserve_functions.R
\name{get_ns_state_list}
\alias{get_ns_state_list}
\title{This function exports all species from a state with NS state rank.}
\usage{
get_ns_state_list(state, taxonomy = TRUE)
}
\arguments{
\item{state}{A character with state code, exmaple: "CO".}

\item{taxonomy}{TRUE of FALSE.  TRUE if you want to return the data with
taxonomic information included.  This can take a long
time for states with large lists.}
}
\value{
A \code{\link[=list]{list()}}
}
\description{
This function exports all species from a state with NS state rank.
}
\examples{
## Not run:
library("mpsgSE")
get_ns_state_list("CO")
## End(Not run)                     
}
