% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seinet.R
\name{get_seinet}
\alias{get_seinet}
\title{Read SEINet data into R}
\usage{
get_seinet(dir_path, crs = NULL)
}
\arguments{
\item{dir_path}{Path to SEINet data folder.}

\item{crs}{Target coordinate reference system (CRS). Either and
\code{sf::st_crs()} object or accepted input string (e.g. "NAD83").
See \code{sf::st_crs()} for more details. Default is NULL. If NULL,
resulting sf object CRS will be WGS84.}
}
\description{
This function reads SEINet data in Darwin Core Archive format into R.
}
\details{
Additional details...
}
\examples{
## Not run:

devtools::install_github("fs-mschmidty/mpsgSE")
library("mpsgSE")

# Path to data folder
data_folder <- file.path("T:/path/to/project/directory"/data/SEINet)

# Pull data from existing SEINet query
sei_dat <- get_seinet(dat_dir = data_folder, crs = "NAD83")

## End(Not run)                     
}
