% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_synonyms.R
\name{get_synonyms}
\alias{get_synonyms}
\title{This function takes in a dataframe of species with assigned taxon_ids and outputs a list of potential synonyms from the GBIF Backbone taxonomy.  If more than one species has a taxon_id assigned to it it will only returns synonyms for one species. This returns many fields, however only the cananical name is used.}
\usage{
get_synonyms(x)
}
\arguments{
\item{x}{a vector of unique taxon_ids from get_taxonomies}
}
\description{
This function takes in a dataframe of species with assigned taxon_ids and outputs a list of potential synonyms from the GBIF Backbone taxonomy.  If more than one species has a taxon_id assigned to it it will only returns synonyms for one species. This returns many fields, however only the cananical name is used.
}
