% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxonomies.R
\name{get_taxonomies}
\alias{get_taxonomies}
\title{Get Taxonomies from GBIF}
\usage{
get_taxonomies(spp_list, query_field = "scientific_name")
}
\arguments{
\item{spp_list}{A data frame containing valid scientific species names.}

\item{query_field}{The name of the variable with valid scientific names.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
This function adds taxonomy information from GBIF to any data frame that has
valid scientific names and returns a tibble. \code{gbif_taxonID} is the GBIF
ID for the given scientific name and full taxonomy from the GBIF backbone
taxonomies database. \code{taxon_id} is ID number of the accepted taxonomy
from the GBIF backbone.
}
\examples{
## Not run:

library(mpsgSE)
spp_list <- get_taxonomies(sp_list_ex)

## End(Not run)                     
}
