% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_swap_list.R
\docType{data}
\name{id_swap}
\alias{id_swap}
\title{Idaho SWAP List}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 282 rows and 21 columns.
}
\usage{
id_swap
}
\description{
\code{id_swap} Idaho SWAP list from 2023.
}
\keyword{datasets}
