% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imbcr.R
\docType{data}
\name{imbcr_mgmt_units}
\alias{imbcr_mgmt_units}
\title{List of IMBCR Management Units on National Forests and Grasslands}
\format{
\subsection{\code{imbcr_mgmt_units}}{
}
}
\usage{
imbcr_mgmt_units
}
\description{
List of IMBCR Management Units on National Forests and Grasslands
}
\keyword{datasets}
