% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imbcr.R
\name{imbcr_mgmt_units}
\alias{imbcr_mgmt_units}
\title{View IMBCR Management Units}
\usage{
imbcr_mgmt_units(file_path = NULL)
}
\arguments{
\item{file_path}{Path to *.rda file. Default is NULL. IF NULL theis function
reads the data from the MPSG T-drive folder.}
}
\value{
A vector.
}
\description{
This function lists the IMBCR Forest Service where the IMBCR has collected
data.
}
\details{
You must be connected to the T-drive, e.g., through VPN, to use this function
unless you have the *.rda file on your local machine.
}
\examples{
## Not run:

library("mpsgSE")

imbcr_mgmt_units()

## End(Not run)                     
}
