% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mt_swap}
\alias{mt_swap}
\title{2023 Montana SWAP List}
\format{
A data frame of 154 observations and 12 variables.
\describe{
\item{group}{Taxonomic group of species.}
\item{scientific_name}{Scientific name of species.}
\item{common_name}{Common name of species.}
\item{gbif_taxonID}{GBIF taxon ID}
\item{kingdom}{Taxonomic Kingdom}
\item{phylum}{Taxonomic Phylum}
\item{class}{Taxonomic Class}
\item{order}{Taxonomic Order}
\item{family}{Taxonomic Family}
\item{genus}{Taxonomic Genus}
\item{species}{Taxonomic Species}
\item{taxon_id}{MPSG taxon ID.}
}
}
\source{
description
}
\usage{
mt_swap
}
\description{
2023 Montana SWAP List
}
\keyword{datasets}
