% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{read_edw_lyr}
\alias{read_edw_lyr}
\title{Read spatial data from the Forest Service ArcGIS REST Services directory}
\usage{
read_edw_lyr(map_name, layer = 0, crs = "EPSG:26912")
}
\arguments{
\item{map_name}{Character. Name of map layer.}

\item{layer}{Integer. Number of layer to read. Default is  zero (0).}

\item{crs}{Coordinate reference system (crs). Default is EPSG:26912 (NAD83
UTM Zone 12).}
}
\value{
An \link{sf} object or raster.
}
\description{
This function reads features from the Forest Service ArcGIS REST Services
Directory, https://apps.fs.usda.gov/arcx/rest/services/EDW, using the
\link{arcgislayers} package.
}
\examples{
library(mpsgSE)

# Administrative Boundary for the Dixie National Forest
admin_bndry <- read_edw_lyr("EDW_ForestSystemBoundaries_01", layer = 1) |> 
  dplyr::filter(forestname == "Dixie National Forest")

}
