% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfss.R
\docType{data}
\name{rfss}
\alias{rfss}
\title{Regional Forester's Sensitive Species List for Regions 1-10}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4152 rows and 11 columns.
}
\usage{
rfss
}
\description{
\code{rfss} is a comprehensive master list from 2024.
\code{rfss_taxonomy} is the taxonomic classifications returned form GBIF.
}
\keyword{datasets}
